1)создал абстрактный класс карты 2)функция load_image теперь в отдельном файле
